Support for "User side of Fontnames"
Until now, PDFBox always suggested to have the machine side of font
handling. We do have to implement a RTF to PDF parser, which needs the
counter part. Word offers the English names of the font names. So there
was the need to learn the font mapper to find Arial Bold instead of
ArialMTBold

The new function getTrueTypeFontOfFamily gives the Fontdescriptor
precedence and add the estimated name parts to the font family name.
e.g. Arial with style bold searches for a Font with the english name
"Arial Bold".